Parallelize the discovery activies
Both cloud and local discovery can take some time,
which if done serially could result in timeouts
that would exceed the external recorder tolerance.
To mitigate against that, some of the timeouts were
set rather short, which could sometimes result in
not discovering all available tuners.  This patch
runs the discovery in parallel allowing the various
individual step timeouts to be increased, and makes
additional attempts to discover tuners.