use dns.resolver to test dns resolution before python requests.
Adds new requirement of dnspython (python3-dns package for some
distros).

dns resolution timeouts are not configurable with python requests,
which can result in excessive wait.  So we will do a preliminary
test to see if resolution can be accomplished in a timely manor,
and if not, abort early, as accessing the cloud for tuner discovery
is usually optional.  There is no assurance that the pre-test will
not result in an excessive dns lookup time by requests, but it does
improve the odds.

This should mitigate against a very specific (and likely extremely
rare) identified case where the path to the authoritative dns
server for silicondust.net) had extremely high latency and loss but
the dns resolver kept trying for quite some time.