prioritize atsc3 channels for tuning order
add in a heuristic that will prioritize atsc3
channels on a tuner and de-prioritize the
atsc1 channels on that tuner, such that for
mixed tuner devices it will (more likely)
be used primarily for atsc3 channel requests
as atsc1 channels will be more likely to be
selected first.

This heuristic will not work well if one
has multiple mixed mode tuners as the tuners
themselves will currently fallback to an
atsc1 tuner if a streaming request is made
and all the atsc3 tuners are in use (there
is currently no way to indicate to a mixed
mode tuner to not fallback).