Add support for ExternalRecorder channelscan
Implemented the LoadChannels, FirstChannel and NextChannel commands of
the ExternalRecorder protocol so that the ExternalRecorder can do a channel scan.
The myththdhrrecorder did already read the channel lineup from the HDHomeRun but previously,
to get the channel information in MythTV it needed to be entered manually with mythtv-setup.
The code is tested with the HDHomeRun HDHR3-4DC for DVB-C and with the HDHR5-2DT for DVB-T2.
The ExternalRecorder asks for a callsign but there is no callsign available
in the channel lineup so the channel name is used here as well.
There will be some changes needed to support not only DVB but also ATSC because
of the major_minor channel number and the callsign.